core 1.消息通知添加extend字段（记录详情） 2.消息titel补充 app 1.收藏价格显示bug fix

+++ litemall-wx-api/src/main/java/org/linlinjava/litemall/wx/service/WxCollectService.java
@@ -70,8 +70,10 @@ public class WxCollectService {
      */
     public Object delete(Integer userId, Integer goodsId) {
         LitemallCollect litemallCollect = collectService.queryByGoodsId(userId, goodsId);
-        collectService.deleteByGoodsId(userId, goodsId);
-        collectAccessoryService.deleteByCollectId(litemallCollect.getId());
+        if(litemallCollect != null){
+            collectService.deleteByGoodsId(userId, goodsId);
+            collectAccessoryService.deleteByCollectId(litemallCollect.getId());
+        }
         return ResponseUtil.ok();
     }
 }
