+++ litemall-wx-api/src/main/java/org/linlinjava/litemall/wx/web/WxMsgController.java
@@ -5,14 +5,13 @@ import org.apache.commons.logging.LogFactory;
 import org.linlinjava.litemall.core.util.ResponseUtil;
 import org.linlinjava.litemall.core.validator.Order;
 import org.linlinjava.litemall.core.validator.Sort;
-import org.linlinjava.litemall.db.domain.LitemallMsg;
-import org.linlinjava.litemall.db.service.LitemallMsgService;
+import org.linlinjava.litemall.db.domain.LitemallNotice;
+import org.linlinjava.litemall.db.domain.LitemallNoticeWithBLOBs;
+import org.linlinjava.litemall.db.service.LitemallNoticeService;
 import org.linlinjava.litemall.wx.annotation.LogAnno;
 import org.linlinjava.litemall.wx.annotation.LoginUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import javax.validation.constraints.NotNull;
 
 /**
  * @author ：stephen
@@ -26,7 +25,7 @@ public class WxMsgController {
     private final Log logger = LogFactory.getLog(WxMsgController.class);
 
     @Autowired
-    private LitemallMsgService litemallMsgService;
+    private LitemallNoticeService litemallMsgService;
 
     /**
      * 消息列表
@@ -35,10 +34,13 @@ public class WxMsgController {
     @LogAnno
     public Object list(@LoginUser Integer userId,
                        @RequestParam(defaultValue = "1") Integer page,
-                       @RequestParam(defaultValue = "10") Integer size,
-                       @Sort @RequestParam(defaultValue = "mark_read,add_time") String sort,
-                       @Order @RequestParam(defaultValue = "asc") String order){
-        return ResponseUtil.okList(litemallMsgService.querySelective(userId, page, size, sort, order));
+                       @RequestParam(defaultValue = "10") Integer limit,
+                       @Sort @RequestParam(defaultValue = "add_time") String sort,
+                       @Order @RequestParam(defaultValue = "desc") String order){
+        if(userId == null){
+            return ResponseUtil.unlogin();
+        }
+        return ResponseUtil.okList(litemallMsgService.querySelective(userId, page, limit, sort, order));
     }
 
     /**
@@ -47,7 +49,10 @@ public class WxMsgController {
     @PostMapping("/mark/{id}")
     @LogAnno
     public Object markRead(@LoginUser Integer userId, @PathVariable Integer id){
-        LitemallMsg msg = new LitemallMsg();
+        if(userId == null){
+            return ResponseUtil.unlogin();
+        }
+        LitemallNoticeWithBLOBs msg = new LitemallNoticeWithBLOBs();
         msg.setId(id);
         msg.setMarkRead(true);
         litemallMsgService.updateById(msg);
@@ -60,7 +65,10 @@ public class WxMsgController {
     @PostMapping("/mark/all")
     @LogAnno
     public Object markAll(@LoginUser Integer userId){
-        LitemallMsg msg = new LitemallMsg();
+        if(userId == null){
+            return ResponseUtil.unlogin();
+        }
+        LitemallNotice msg = new LitemallNotice();
         msg.setUserId(userId);
         msg.setMarkRead(true);
         return ResponseUtil.ok(litemallMsgService.updateMarkByUserId(msg));
@@ -71,7 +79,10 @@ public class WxMsgController {
      */
     @DeleteMapping("/delete/{id}")
     @LogAnno
-    public Object delete(@LoginUser Integer uerId, @PathVariable Integer id){
+    public Object delete(@LoginUser Integer userId, @PathVariable Integer id){
+        if(userId == null){
+            return ResponseUtil.unlogin();
+        }
         litemallMsgService.deleteById(id);
         return ResponseUtil.ok();
     }
@@ -82,7 +93,10 @@ public class WxMsgController {
     @DeleteMapping("/delete/all")
     @LogAnno
     public Object deleteAll(@LoginUser Integer userId){
-        LitemallMsg msg = new LitemallMsg();
+        if(userId == null){
+            return ResponseUtil.unlogin();
+        }
+        LitemallNotice msg = new LitemallNotice();
         msg.setUserId(userId);
         msg.setMarkRead(true);
         return ResponseUtil.ok(litemallMsgService.deleteByUserId(userId));
